move all functions unter utils.py
