+++ examples/text-generation/run_generation.py
@@ -19,41 +19,16 @@ Conditional text generation on Habana Gaudi/Gaudi2.
 """
 
 import argparse
-import copy
 import json
 import logging
-import os
-import shutil
-import tempfile
 import time
 from pathlib import Path
 
 import torch
-from transformers import AutoConfig, AutoModelForCausalLM, AutoTokenizer
-from transformers.utils import check_min_version
-from utils import count_hpu_graphs, override_prints, setup_distributed
+from utils import count_hpu_graphs, initialize_model
 
-from optimum.habana.checkpoint_utils import (
-    get_ds_injection_policy,
-    get_repo_root,
-    model_is_optimized,
-    model_on_meta,
-    write_checkpoints_json,
-)
-from optimum.habana.utils import get_hpu_memory_stats, set_seed
+from optimum.habana.utils import get_hpu_memory_stats
 
-
-try:
-    from optimum.habana.utils import check_optimum_habana_min_version
-except ImportError:
-
-    def check_optimum_habana_min_version(*a, **b):
-        return ()
-
-
-# Will error if the minimal version of Transformers is not installed. Remove at your own risks.
-check_min_version("4.34.0")
-check_optimum_habana_min_version("1.9.0.dev0")
 
 logging.basicConfig(
     format="%(asctime)s - %(levelname)s - %(name)s - %(message)s",
@@ -61,58 +36,6 @@ logging.basicConfig(
     level=logging.INFO,
 )
 logger = logging.getLogger(__name__)
-
-
-def setup_env(args):
-    if args.global_rank == 0:
-        os.environ.setdefault("GRAPH_VISUALIZATION", "true")
-        shutil.rmtree(".graph_dumps", ignore_errors=True)
-
-    if args.world_size > 0:
-        os.environ.setdefault("PT_HPU_LAZY_ACC_PAR_MODE", "0")
-        os.environ.setdefault("PT_HPU_ENABLE_LAZY_COLLECTIVES", "true")
-
-    # Tweak generation so that it runs faster on Gaudi
-    from optimum.habana.transformers.modeling_utils import adapt_transformers_to_gaudi
-
-    adapt_transformers_to_gaudi()
-
-
-def setup_device(args):
-    return torch.device(args.device)
-
-
-def initialize_model(args):
-    init_start = time.perf_counter()
-    setup_distributed(args)
-    override_prints(args.global_rank == 0 or args.verbose_workers, logger)
-    setup_env(args)
-    setup_device(args)
-    set_seed(args.seed)
-    get_repo_root(args.model_name_or_path, local_rank=args.local_rank, token=args.token)
-    use_deepspeed = args.world_size > 0
-    if use_deepspeed or args.bf16:
-        model_dtype = torch.bfloat16
-    else:
-        model_dtype = torch.float
-        args.attn_softmax_bf16 = False
-
-    model_kwargs = {
-        "revision": args.model_revision,
-        "token": args.token,
-    }
-    model = (
-        setup_model(args, model_dtype, model_kwargs)
-        if not use_deepspeed
-        else setup_distributed_model(args, model_dtype, model_kwargs)
-    )
-    tokenizer, model = setup_tokenizer(args, model)
-    generation_config = setup_generation_config(args, model, tokenizer)
-    init_end = time.perf_counter()
-    logger.info(f"Args: {args}")
-    logger.info(f"device: {args.device}, n_hpu: {args.world_size}, bf16: {model_dtype == torch.bfloat16}")
-    logger.info(f"Model initialization took {(init_end - init_start):.3f}s")
-    return model, tokenizer, generation_config
 
 
 def setup_parser(parser):
@@ -286,126 +209,10 @@ def setup_parser(parser):
     return args
 
 
-def setup_model(args, model_dtype, model_kwargs):
-    logger.info("Single-device run.")
-    model = AutoModelForCausalLM.from_pretrained(args.model_name_or_path, torch_dtype=model_dtype, **model_kwargs)
-    model = model.eval().to(args.device)
-    model = peft_model(args, model, model_dtype)
-    if args.use_hpu_graphs:
-        from habana_frameworks.torch.hpu import wrap_in_hpu_graph
-
-        model = wrap_in_hpu_graph(model)
-    return model
-
-
-def setup_tokenizer(args, model):
-    tokenizer_kwargs = {
-        "revision": args.model_revision,
-        "token": args.token,
-    }
-    if args.bad_words is not None or args.force_words is not None:
-        tokenizer_kwargs["add_prefix_space"] = True
-    tokenizer = AutoTokenizer.from_pretrained(args.model_name_or_path, **tokenizer_kwargs)
-    if not model.config.is_encoder_decoder:
-        tokenizer.padding_side = "left"
-    # Some models like GPT2 do not have a PAD token so we have to set it if necessary
-    if model.config.model_type == "llama":
-        # unwind broken decapoda-research config
-        model.generation_config.pad_token_id = 0
-        model.generation_config.bos_token_id = 1
-        model.generation_config.eos_token_id = 2
-        tokenizer.bos_token_id = model.generation_config.bos_token_id
-        tokenizer.eos_token_id = model.generation_config.eos_token_id
-        tokenizer.pad_token_id = model.generation_config.pad_token_id
-        tokenizer.pad_token = tokenizer.decode(tokenizer.pad_token_id)
-        tokenizer.eos_token = tokenizer.decode(tokenizer.eos_token_id)
-        tokenizer.bos_token = tokenizer.decode(tokenizer.bos_token_id)
-
-    if tokenizer.pad_token is None:
-        tokenizer.pad_token = tokenizer.eos_token
-        model.generation_config.pad_token_id = model.generation_config.eos_token_id
-    return tokenizer, model
-
-
-def setup_distributed_model(args, model_dtype, model_kwargs):
-    import deepspeed
-
-    logger.info("DeepSpeed is enabled.")
-    config = AutoConfig.from_pretrained(args.model_name_or_path, **model_kwargs)
-    load_to_meta = model_on_meta(config)
-
-    if load_to_meta:
-        # Construct model with fake meta tensors, later will be replaced on devices during ds-inference ckpt load
-        with deepspeed.OnDevice(dtype=model_dtype, device="meta"):
-            model = AutoModelForCausalLM.from_config(config, torch_dtype=model_dtype)
-        # model loaded to meta is managed differently
-        checkpoints_json = tempfile.NamedTemporaryFile(suffix=".json", mode="+w")
-        write_checkpoints_json(args.model_name_or_path, args.local_rank, checkpoints_json, token=args.token)
-    else:
-        # TODO: revisit placement on CPU when auto-injection is possible
-        with deepspeed.OnDevice(dtype=model_dtype, device="cpu"):
-            model = AutoModelForCausalLM.from_pretrained(
-                args.model_name_or_path, torch_dtype=model_dtype, **model_kwargs
-            )
-    model = model.eval()
-    model = peft_model(args, model, model_dtype)
-    # Initialize the model
-    ds_inference_kwargs = {"dtype": model_dtype}
-    ds_inference_kwargs["tensor_parallel"] = {"tp_size": args.world_size}
-    ds_inference_kwargs["enable_cuda_graph"] = args.use_hpu_graphs
-    ds_inference_kwargs["injection_policy"] = get_ds_injection_policy(config)
-    if load_to_meta:
-        ds_inference_kwargs["checkpoint"] = checkpoints_json.name
-
-    model = deepspeed.init_inference(model, **ds_inference_kwargs)
-    model = model.module
-    return model
-
-
-def setup_generation_config(args, model, tokenizer):
-    bad_words_ids = None
-    force_words_ids = None
-    if args.bad_words is not None:
-        bad_words_ids = [tokenizer.encode(bad_word, add_special_tokens=False) for bad_word in args.bad_words]
-    if args.force_words is not None:
-        force_words_ids = [tokenizer.encode(force_word, add_special_tokens=False) for force_word in args.force_words]
-
-    is_optimized = model_is_optimized(model.config)
-    # Generation configuration
-    generation_config = copy.deepcopy(model.generation_config)
-    generation_config.max_new_tokens = args.max_new_tokens
-    generation_config.use_cache = args.use_kv_cache
-    generation_config.static_shapes = is_optimized
-    generation_config.bucket_size = args.bucket_size if is_optimized else -1
-    generation_config.do_sample = args.do_sample
-    generation_config.num_beams = args.num_beams
-    generation_config.bad_words_ids = bad_words_ids
-    generation_config.force_words_ids = force_words_ids
-    generation_config.num_return_sequences = args.num_return_sequences
-    generation_config.trim_logits = args.trim_logits
-    generation_config.attn_softmax_bf16 = args.attn_softmax_bf16
-    generation_config.limit_hpu_graphs = args.limit_hpu_graphs
-    generation_config.reuse_cache = args.reuse_cache
-    return generation_config
-
-
-def peft_model(args, model, model_dtype):
-    if args.peft_model:
-        import importlib.util
-
-        if importlib.util.find_spec("peft") is None:
-            raise ImportError("The `peft` package is not installed, please run: `pip install peft`.")
-        from peft import PeftModel
-
-        model = PeftModel.from_pretrained(model, args.peft_model)
-        model = model.merge_and_unload().eval().to(model_dtype)
-    return model
-
-
 def main():
     parser = argparse.ArgumentParser()
     args = setup_parser(parser)
-    model, tokenizer, generation_config = initialize_model(args)
+    model, tokenizer, generation_config = initialize_model(args, logger)
 
     import habana_frameworks.torch.hpu as torch_hpu
